Finally got ourself a "working" m16c environment, at least the hello world example is working atm. Probably a few bugs lurking out there but for now it seems to be a good foundation to stand upon.

+++ env/m16c/types.h
@@ -61,7 +61,7 @@ typedef m16c_context_t env_context_t;
  * Should evaluate to the type that is used to represent time within the
  * environment.
  */
-typedef int env_time_t;
+typedef unsigned long env_time_t;
 
 /**
  * \brief M16C Environment result typdef.
